Split hunk handlers into separate files.
