+++ tools/hunk/hunk_end.go
@@ -1,0 +1,7 @@
+package hunk
+
+type HunkEnd struct{}
+
+func (h HunkEnd) Name() string {
+	return "HUNK_END"
+}
